Add files via upload
Still facing issues with the training.  It looks like exploration was limited and learning rate was a tad high (it quickly set preferred value to always be the third column).  Also, looking at the neural net initial prediction, mcts policy gradient collapsed to zeros and a singular one way too quickly, meaning exploration was no where close to where it should be.